intan-rhx: Wire up time sync code
The code was written to require minimal changes on the old RHX2000 code
(so we could share some code with Intan) - turns out it works quite well
for the new module as well, and can even be simplified in future!

+++ modules/intan-rhx/Engine/API/Hardware/rhxglobals.h
@@ -265,7 +265,7 @@ const int Register59MISOA = 53;
 const int Register59MISOB = 58;
 
 const int BytesPerWord = 2;
-const int BytesPerSyTimestamp = sizeof(uint32_t);
+const int BytesPerSyTimestamp = sizeof(uint64_t);
 
 // Trigonometric constants
 const double Pi = 3.14159265359;
